# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.AppLogsResponse:
    properties:
      logs:
        type: string
    type: object
  handler.CreateAppRequest:
    properties:
      image:
        example: nginx:latest
        maxLength: 256
        type: string
      name:
        example: my-nginx
        maxLength: 63
        minLength: 1
        type: string
      port:
        example: 80
        maximum: 65535
        minimum: 1
        type: integer
      replicas:
        example: 2
        maximum: 10
        minimum: 1
        type: integer
    required:
    - image
    - name
    - replicas
    type: object
  handler.CreateProjectRequest:
    properties:
      name:
        example: 个人网站
        maxLength: 64
        type: string
    required:
    - name
    type: object
  handler.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  handler.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handler.OAuth2LoginResponse:
    properties:
      auth_url:
        example: https://auth.example.com/application/o/authorize/?client_id=astro
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - password
    - username
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Astro 容器即服务平台 API 文档
  title: Astro API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录获取 Token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 用户登录
      tags:
      - 用户
  /oauth2/{provider}/callback:
    get:
      description: 使用授权码换取本系统 JWT
      parameters:
      - description: 公开 Provider 别名
        example: bytcloudauth
        in: path
        name: provider
        required: true
        type: string
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态参数
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/handler.Response'
      summary: OAuth2 回调
      tags:
      - 用户
  /oauth2/{provider}/login:
    get:
      description: 获取 OAuth2/OIDC Provider 授权 URL
      parameters:
      - description: 公开 Provider 别名
        example: bytcloudauth
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.OAuth2LoginResponse'
              type: object
        "400":
          description: Provider 不可用
          schema:
            $ref: '#/definitions/handler.Response'
      summary: OAuth2 登录入口
      tags:
      - 用户
  /projects:
    get:
      description: 获取当前用户的所有项目
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 获取项目列表
      tags:
      - 项目
    post:
      consumes:
      - application/json
      description: 创建项目并建立独立的 Kubernetes Namespace
      parameters:
      - description: 项目信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 创建项目
      tags:
      - 项目
  /projects/{project_id}:
    delete:
      description: 删除空项目及其 Kubernetes Namespace
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 删除项目
      tags:
      - 项目
    get:
      description: 获取当前用户的指定项目
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 获取项目详情
      tags:
      - 项目
  /projects/{project_id}/apps:
    get:
      description: 获取指定项目的所有应用
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 获取应用列表
      tags:
      - 应用
    post:
      consumes:
      - application/json
      description: 在指定项目中创建一个新的容器应用
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 创建应用
      tags:
      - 应用
  /projects/{project_id}/apps/{id}:
    delete:
      description: 删除指定项目内的应用
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 应用不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 删除应用
      tags:
      - 应用
    get:
      description: 获取指定项目内应用的详细信息
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 应用不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 获取应用详情
      tags:
      - 应用
  /projects/{project_id}/apps/{id}/logs:
    get:
      description: 获取指定项目内应用的容器日志
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      - default: 100
        description: 日志行数
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.AppLogsResponse'
              type: object
      security:
      - Bearer: []
      summary: 获取应用日志
      tags:
      - 应用
  /projects/{project_id}/apps/{id}/restart:
    post:
      description: 重启指定项目内的应用
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 重启成功
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 重启应用
      tags:
      - 应用
  /projects/{project_id}/apps/{id}/start:
    post:
      description: 启动指定项目内的应用
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 启动成功
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 启动应用
      tags:
      - 应用
  /projects/{project_id}/apps/{id}/stop:
    post:
      description: 停止指定项目内的应用
      parameters:
      - description: 项目ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 停止成功
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 停止应用
      tags:
      - 应用
  /register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 用户注册
      tags:
      - 用户
securityDefinitions:
  Bearer:
    description: 请输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"